refactor(features): update signin/signup logic
Move out logic for extracting data to `helpers/database.ts`
